todo service

+++ ToDoService/src/main/java/com/app/todoservice/repository/ItemDetailsRepo.java
@@ -1,8 +1,7 @@
 package com.app.todoservice.repository;
 
 import com.app.todoservice.entity.ItemDetails;
-import com.app.todoservice.entity.Items;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemDetailsRepo extends JpaRepository<ItemDetails, Integer> {
+public interface ItemDetailsRepo extends JpaRepository<ItemDetails, Long> {
 }
